[v1.x] Move RequestBodyLimitMiddleware out of the Streamable HTTP manager module
The middleware and DEFAULT_MAX_REQUEST_BODY_SIZE are now used by the SSE
transport and the OAuth routes as well, so they move next to the other
shared HTTP request checks in mcp.server.transport_security. Both names
remain importable from mcp.server.streamable_http_manager (listed in its
__all__). The middleware's own unit tests move with it, plus the
mid-stream disconnect replay case main already carries; no behaviour
change.
